requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.1.7
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
pillow==11.1.0
sqlparse==0.5.3
uvicorn==0.34.0
whitenoise==6.9.0
psycopg==3.2.5
psycopg2==2.9.10
roman==5.0
dj-database-url==2.3.0
cloudinary==1.43.0
django-cloudinary-storage==0.3.0
python-dotenv==1.0.1
djangorestframework==3.15.2